Change type of timeout interface from float to unsigned int: fix #269

+++ jubatus/client/classifier_client.hpp
@@ -17,7 +17,7 @@ namespace client {
 
 class classifier {
  public:
-  classifier(const std::string& host, uint64_t port, double timeout_sec)
+  classifier(const std::string& host, uint64_t port, unsigned int timeout_sec)
       : c_(host, port) {
     c_.set_timeout(timeout_sec);
   }
