+++ jubatus/client/nearest_neighbor_client.hpp
@@ -17,7 +17,8 @@ namespace client {
 
 class nearest_neighbor {
  public:
-  nearest_neighbor(const std::string& host, uint64_t port, double timeout_sec)
+  nearest_neighbor(const std::string& host, uint64_t port,
+       unsigned int timeout_sec)
       : c_(host, port) {
     c_.set_timeout(timeout_sec);
   }
